Служебные функции для получения списков рецептов по категориям и их сортировки вынесены в отдельный файл.
